Fixing some structural errors in the AST
The AST had some errors, these were mainly surrounding the
hierarchy of the nodes. The main one of these was that the Expr
node needed to extends that of Rvalue,  which it did not. There is
now a distinct hierachy within the AST, with a general type
ASTnode at the top. The various types that extends more than one
supertype are now also ensured to extend both.

Some namings were also fixed to reflect the fact that some types
could be extensions of multiple other types.  For instance,
ExprRight, which was the type Expr used in an Rvalue, is simply
Expr, as there is no distinction between them and an Expr is an
Rvalue.

ParserBridges have also been fixed although some may require a
type change for their second argument, this  will be an easy fix
though which we can make if we need.